[Cleanup] Renamed Assign/Dismiss command names for clarity
