+++ src/main/java/seedu/address/logic/commands/DismissCommand.java
@@ -16,25 +16,25 @@ import seedu.address.model.task.ReadOnlyTask;
 import seedu.address.model.task.exceptions.DuplicateTaskException;
 import seedu.address.model.task.exceptions.TaskNotFoundException;
 
-/** Assigns at least 1 person to a specified task in the Address Book**/
-public class AssignTaskCommand extends Command {
-    public static final String COMMAND_WORD = "assign";
-    public static final String COMMAND_ALIAS = "as";
+/** Dismisses at least 1 person from a specified task in the Address Book**/
+public class DismissCommand extends Command {
+    public static final String COMMAND_WORD = "dismiss";
+    public static final String COMMAND_ALIAS = "ds";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Assigns people to a task in the Address Book. "
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Dismisses people from a task in the Address Book. "
             + "Parameters: "
             + "PERSON INDEXES... "
             + PREFIX_TARGET + "TASK ";
 
-    public static final String MESSAGE_SUCCESS = "Assigned %1$s people to \n%2$s";
+    public static final String MESSAGE_SUCCESS = "Dismissed %1$s people from task \n%2$s";
     public static final String MESSAGE_INVALID_TARGET_ARGS = "Only 1 task index should be specified";
     public static final String MESSAGE_INVALID_PERSONS_ARGS = "At least 1 person index must be specified";
-    public static final String MESSAGE_NONE_ASSIGNED = "All the specified persons are already assigned to this task";
+    public static final String MESSAGE_NONE_ASSIGNED = "None of the specified persons are assigned to this task";
 
     private ArrayList<Index> personIndexes;
     private Index taskIndex;
 
-    public AssignTaskCommand(ArrayList<Index> personIndexes, Index taskIndex) {
+    public DismissCommand(ArrayList<Index> personIndexes, Index taskIndex) {
         assert(personIndexes.size() > 0);
         this.personIndexes = personIndexes;
         this.taskIndex = taskIndex;
@@ -43,50 +43,50 @@ public class AssignTaskCommand extends Command {
     @Override
     public CommandResult execute() throws CommandException {
         List<ReadOnlyTask> tasksList = model.getFilteredTaskList();
-        ArrayList<ReadOnlyPerson> personIndexes = createPersonsToAssign(this.personIndexes);
+        ArrayList<ReadOnlyPerson> personIndexes = createPersonsToDismiss(this.personIndexes);
 
         if (taskIndex.getZeroBased() >= tasksList.size()) {
             throw new CommandException(Messages.MESSAGE_INVALID_TASK_DISPLAYED_INDEX);
         }
-        ReadOnlyTask assignedTask = tasksList.get(taskIndex.getZeroBased());
+        ReadOnlyTask dismissedTask = tasksList.get(taskIndex.getZeroBased());
         try {
-            model.assignToTask(personIndexes, assignedTask);
+            model.dismissFromTask(personIndexes, dismissedTask);
         } catch (TaskNotFoundException tnfe) {
             throw new AssertionError("The specified task cannot be missing");
         } catch (DuplicateTaskException dte) {
             throw new CommandException(MESSAGE_NONE_ASSIGNED);
         }
-        return new CommandResult(String.format(MESSAGE_SUCCESS, personIndexes.size(), assignedTask));
+        return new CommandResult(String.format(MESSAGE_SUCCESS, personIndexes.size(), dismissedTask));
     }
 
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof AssignTaskCommand // instanceof handles nulls
-                && this.personIndexes.equals(((AssignTaskCommand) other).personIndexes)
-                && this.taskIndex.equals(((AssignTaskCommand) other).taskIndex)); // state check
+                || (other instanceof DismissCommand // instanceof handles nulls
+                && this.personIndexes.equals(((DismissCommand) other).personIndexes)
+                && this.taskIndex.equals(((DismissCommand) other).taskIndex)); // state check
     }
 
     /**
      * Creates a {@code ArrayList} that contains all the {@code ReadOnlyPerson} converted from the {@Code Index}
      * @throws CommandException if the specified Index is out of range
      */
-    public ArrayList<ReadOnlyPerson> createPersonsToAssign (ArrayList<Index> indexes)  throws CommandException {
+    public ArrayList<ReadOnlyPerson> createPersonsToDismiss (ArrayList<Index> indexes)  throws CommandException {
         HashSet<ReadOnlyPerson> addedPersons = new HashSet<>();
-        ArrayList<ReadOnlyPerson> personsToAssign = new ArrayList<>();
+        ArrayList<ReadOnlyPerson> personsToDismiss = new ArrayList<>();
         List<ReadOnlyPerson> personsList = model.getFilteredPersonList();
         try {
             for (Index i : personIndexes) {
-                ReadOnlyPerson toAssign = personsList.get(i.getZeroBased());
-                if (!addedPersons.contains(toAssign)) {
-                    addedPersons.add(toAssign);
-                    personsToAssign.add(toAssign);
+                ReadOnlyPerson toDismiss = personsList.get(i.getZeroBased());
+                if (!addedPersons.contains(toDismiss)) {
+                    addedPersons.add(toDismiss);
+                    personsToDismiss.add(toDismiss);
                 }
             }
         } catch (IndexOutOfBoundsException e) {
             throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
         }
 
-        return personsToAssign;
+        return personsToDismiss;
     }
 }
