+++ src/main/java/seedu/address/logic/parser/DismissTaskCommandParser.java
@@ -9,32 +9,32 @@ import static seedu.address.logic.parser.ParserUtil.MESSAGE_INVALID_INDEX;
 import java.util.ArrayList;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DismissTaskCommand;
+import seedu.address.logic.commands.DismissCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
-/** Parses input arguments, and creates a new {@code DismissTaskCommand} object**/
-public class DismissTaskCommandParser implements Parser<DismissTaskCommand> {
+/** Parses input arguments, and creates a new {@code DismissCommand} object**/
+public class DismissTaskCommandParser implements Parser<DismissCommand> {
     /**
-     * Parses the given {@code String} of arguments in the context of the DismissTaskCommand
-     * and returns an DismissTaskCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the DismissCommand
+     * and returns an DismissCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public DismissTaskCommand parse(String args) throws ParseException {
+    public DismissCommand parse(String args) throws ParseException {
         ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args, PREFIX_FROM);
         if (!argMultimap.getValue(PREFIX_FROM).isPresent()) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DismissTaskCommand.MESSAGE_USAGE));
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DismissCommand.MESSAGE_USAGE));
         }
         String target = argMultimap.getValue(PREFIX_FROM).get();
         String persons = argMultimap.getPreamble();
         ArrayList<Index> targetIndexes = parseIndexes(target);
         ArrayList<Index> personIndexes = parseIndexes(persons);
         if (targetIndexes.size() != 1) {
-            throw new ParseException(DismissTaskCommand.MESSAGE_INVALID_TARGET_ARGS);
+            throw new ParseException(DismissCommand.MESSAGE_INVALID_TARGET_ARGS);
         } else if (personIndexes.size() < 1) {
-            throw new ParseException(DismissTaskCommand.MESSAGE_INVALID_PERSONS_ARGS);
+            throw new ParseException(DismissCommand.MESSAGE_INVALID_PERSONS_ARGS);
         }
         Index taskIndex = targetIndexes.get(0);
-        return new DismissTaskCommand(personIndexes, taskIndex);
+        return new DismissCommand(personIndexes, taskIndex);
     }
 
     /**
